requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.24.0
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.1
protobuf==6.33.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.1.15
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.24
tokenizers==0.22.2
torch==2.11.0.dev20260214+cu128
torchvision==0.26.0.dev20260214+cu128
tqdm==4.67.3
transformers==5.2.0
typer==0.23.1
typer-slim==0.23.1
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.6.0
Werkzeug==3.1.5